% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_function.R
\name{check_coreAnalysis_arguments}
\alias{check_coreAnalysis_arguments}
\title{A helper function for checking coreAnalysis argumnets}
\usage{
check_coreAnalysis_arguments(x, test_statistic = c("z", "t", "F", "chisq",
  "correlation"), pval_threshold = 0.05, threshold = NULL, df1 = NULL,
  df2 = NULL)
}
\description{
A helper function for checking coreAnalysis argumnets
}
