% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq-statistic.R
\name{chisq_conditional_density}
\alias{chisq_conditional_density}
\title{The conditional chi-square density}
\usage{
chisq_conditional_density(ncp, x, df, threshold)
}
\arguments{
\item{ncp}{the non-centrality parameter}

\item{x}{the observed value}

\item{df}{the chi-square degrees of freedom}
}
\description{
The conditional chi-square density
}
