% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq-statistic.R
\name{chisq_conditional_ncp_ci}
\alias{chisq_conditional_ncp_ci}
\title{Conditional test-inversion confidence intervals for chi-square ncps}
\usage{
chisq_conditional_ncp_ci(x, df, pval_threshold = 0.05, threshold = NULL,
  confidence_level = 0.95, normalize = FALSE)
}
\arguments{
\item{x}{the observed value}

\item{df}{the degrees of freedom}

\item{pval_threshold}{the pvalue corresponding to the chisq-statistic (under the null)
must be below this value to be observed.}

\item{threshold}{an alternative way to specify the selection threshold. If specificed,
then the selection rule is assumed to be \code{x > threshold}. Takes precedence to
\code{pval_threshold}.}

\item{confidence_level}{the desired confidence level of the confidence
interval.}

\item{normalize}{whether to report the non-centrality parameter per degree of
freedom}
}
\description{
Conditional test-inversion confidence intervals for chi-square ncps
}
