% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq-statistic.R
\name{chisq_conditional_ncp_mle}
\alias{chisq_conditional_ncp_mle}
\title{Compute the mle of the non-centrality parameter of a chi-square test-statistic conditional on selection}
\usage{
chisq_conditional_ncp_mle(x, df, pval_threshold = 0.05, threshold = NULL,
  normalize = FALSE)
}
\arguments{
\item{x}{the observed value}

\item{df}{the degrees of freedom}

\item{pval_threshold}{the pvalue corresponding to the chisq-statistic (under the null)
must be below this value to be observed.}

\item{threshold}{an alternative way to specify the selection threshold. If specificed,
then the selection rule is assumed to be \code{x > threshold}. Takes precedence to
\code{pval_threshold}.

#' @param normalize whether to report the non-centrality parameter per degree of
freedom}
}
\description{
Compute the mle of the non-centrality parameter of a chi-square test-statistic conditional on selection
}
