% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq-statistic.R
\name{chisq_ncp_mle}
\alias{chisq_ncp_mle}
\title{Compute the mle of the non-centrality parameter for a chi-square test-statistic}
\usage{
chisq_ncp_mle(x, df, normalize = TRUE)
}
\arguments{
\item{x}{the observed value}

\item{df}{the degrees of freedom}

\item{normalize}{whether to report the non-centrality parameter per degree of
freedom}
}
\description{
Compute the mle of the non-centrality parameter for a chi-square test-statistic
}
