% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_function.R
\name{CoReAnalysis}
\alias{CoReAnalysis}
\title{Conditional replication analysis}
\usage{
CoReAnalysis(x, test_statistic = c("z", "t", "F", "chisq", "correlation"),
  pval_threshold = 0.05, threshold = NULL, df1 = NULL, df2 = NULL,
  confidence_level = 0.95)
}
\arguments{
\item{x}{the observed test-statistic.}

\item{test_statistic}{the type of test-statistic}

\item{pval_threshold}{the threshold a pvalue must cross (be lesser than)
in order for the test-statistic to be observed.}

\item{threshold}{an alternative way to specifiy a threshold. If a
z, t, or correlation then selection rule is assumed to be
\code{abs(x) > threshold}. If F or chisq then the selection rule
is assumed to be \code{x > threshold}. Takes precedence to
\code{pval_threshold}.}

\item{df1}{the degrees of freedom for t and chisq statistics. The sample
size for Pearson's correlation. The numerator degrees of freedom for an
F statistic.}

\item{df2}{the denominator degrees of freedom for an F statistic.}

\item{confidence_level}{the desired confidence level for the
confidence interval.}
}
\description{
Computes conditional maximum likelihood estimators and confidence
confidence intervals for test-statistics that were selected based
on their significance.
}
