% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson-correlation.R
\name{corr_to_z}
\alias{corr_to_z}
\title{Converts Pearson's correaltion to z-statstic via Fisher's Transfomration}
\usage{
corr_to_z(corr, samp_size)
}
\arguments{
\item{corr}{the observed correaltion}

\item{samp_size}{the size of the sample the correaltion was computed from}
}
\description{
Converts Pearson's correaltion to z-statstic via Fisher's Transfomration
}
