% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-statistic.R
\name{f_conditional_density}
\alias{f_conditional_density}
\title{The conditional F density}
\usage{
f_conditional_density(ncp, x, df1, df2, threshold)
}
\arguments{
\item{ncp}{the non-centrality parameter}

\item{x}{the observed value}

\item{df2}{the denominator degress of freedom}

\item{df}{the numerator degrees of freedom}
}
\description{
The conditional F density
}
