% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-statistic.R
\name{f_conditional_ncp_mle}
\alias{f_conditional_ncp_mle}
\title{Compute the mle of the non-centrality parameter of an F test-statistic conditional on selection}
\usage{
f_conditional_ncp_mle(x, df1 = 1, df2 = Inf, pval_threshold = 0.05,
  threshold = NULL)
}
\arguments{
\item{x}{the observed value}

\item{df1}{the numerator degrees of freedom}

\item{df2}{the denominator degrees of freedom}

\item{pval_threshold}{the pvalue corresponding to the chisq-statistic (under the null)
must be below this value to be observed.}

\item{threshold}{an alternative way to specify the selection threshold. If specificed,
then the selection rule is assumed to be \code{x > threshold}. Takes precedence to
\code{pval_threshold}.}
}
\description{
Compute the mle of the non-centrality parameter of an F test-statistic conditional on selection
}
