% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{numerical_invert_ci}
\alias{numerical_invert_ci}
\title{A function for inverting tests numerically}
\usage{
numerical_invert_ci(cdf, quantile, init = 0, step_size = 0.1,
  lbound = -Inf, ubound = Inf, ...)
}
\arguments{
\item{cdf}{a cdf function, the first argument of which is the parameter
for which a ci is to be computed.}

\item{quantile}{the ci quantile to be computed.}

\item{init}{a value from which to start the search.}

\item{step_size}{step size for searching for the neighborhood of the correct
ci quantile.}

\item{lbound}{lower bound on the values the parameter can take.}

\item{ubound}{upper bound on the values the parameter can take.}

\item{...}{additional arguments for \code{cdf} function.}
}
\description{
A function for inverting tests numerically
}
