% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{replication_power}
\alias{replication_power}
\title{Power analysis following Conditional Replicability Analysis}
\usage{
replication_power(object, samp_size_ratio = 2^(seq(from = -2, to = 3, by =
  0.5)), plot = FALSE)
}
\arguments{
\item{object}{a \code{\link[CoReAnalysis]{CoReAnalysis}} object.}

\item{samp_size_ratio}{the ratio of the sample size in the replication study to
the sample size in the original study. For example, if the sample,
size in both studies is the same
then the ratio is 1 and if the replication study will have 200 subjects where
the original study had 100 subjects then the ration will be 100.}

\item{plot}{if TRUE, then a \code{\link[ggplot2]{ggplot2}} object will be returned.}
}
\description{
Computes conditional estimates of power to replicate a study based on
a \code{\link[CoReAnalysis]{CoReAnalysis}} fit.
}
