% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-statistic.R
\name{t_conditional_density}
\alias{t_conditional_density}
\title{A function for computing the (negative) conditional t density}
\usage{
t_conditional_density(ncp, x, df = Inf, threshold)
}
\arguments{
\item{x}{the observed value}

\item{df}{the degrees of freedom of \code{x}.}

\item{threshold}{The threshold defines the selection rule, \code{abs(x) > threshold}.}
}
\description{
A function for computing the (negative) conditional t density
}
