% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-statistic.R
\name{t_conditional_ncp_ci}
\alias{t_conditional_ncp_ci}
\title{Compute a conditional ci for a t-statistic non-centraility parameter}
\usage{
t_conditional_ncp_ci(x, df = Inf, pval_threshold, threshold = NULL,
  confidence_level = 0.95)
}
\arguments{
\item{x}{the observed value}

\item{df}{the degrees of freedom of \code{x}.}

\item{pval_threshold}{the pvalue corresponding to the t-statistic (under the null)
must be below this value to be observed.}

\item{threshold}{an alternative way to specify the selection threshold. If specificed,
then the selection rule is assumed to be \code{abs(x) > threshold}. Takes precedence to
\code{pval_threshold}.}

\item{confidence_level}{the desired confidence level of the confidence
interval.}
}
\description{
Compute a conditional ci for a t-statistic non-centraility parameter
}
