% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-statistic.R
\name{t_conditional_ncp_mle}
\alias{t_conditional_ncp_mle}
\title{A function for computing the MLE of a t distribution conditional on selection}
\usage{
t_conditional_ncp_mle(x, df = Inf, pval_threshold, threshold = NULL,
  iterlim = 1000)
}
\arguments{
\item{x}{the observed value}

\item{df}{the degrees of freedom of \code{x}.}

\item{pval_threshold}{the pvalue corresponding to the t-statistic (under the null)
must be below this value to be observed.}

\item{threshold}{an alternative way to specify the selection threshold. If specificed,
then the selection rule is assumed to be \code{abs(x) > threshold}. Takes precedence to
\code{pval_threshold}.}

\item{iterlim}{maximum number of numerical optimization iterations.
See \code{\link[stats]{nlm}} for details.}
}
\description{
A function for computing the MLE of a t distribution conditional on selection
}
