% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-statistic.R
\name{t_ncp_ci}
\alias{t_ncp_ci}
\title{Computes a naive CI for the t non-centrality parameter}
\usage{
t_ncp_ci(x, df = Inf, confidence_level = 0.95)
}
\arguments{
\item{x}{the observed value}

\item{df}{the degrees of freedom of \code{x}.}

\item{confidence_level}{the desired confidence level of the confidence
interval.}
}
\description{
Computes a naive CI for the t non-centrality parameter
}
