% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-statistic.R
\name{t_ncp_mle}
\alias{t_ncp_mle}
\title{A function for estimating the MLE of the ncp of a t distribution}
\usage{
t_ncp_mle(x, df = Inf)
}
\arguments{
\item{x}{the observed value}

\item{df}{the degrees of freedom of \code{x}.}
}
\description{
A function for estimating the MLE of the ncp of a t distribution
}
