% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_function.R
\name{threshold_conversion}
\alias{threshold_conversion}
\title{Pval_threshold to threshold conversion}
\usage{
threshold_conversion(x, test_statistic = c("z", "t", "F", "chisq",
  "correlation"), pval_threshold = 0.05, threshold = NULL, df1 = NULL,
  df2 = NULL)
}
\arguments{
\item{x}{the observed test-statistic.}

\item{test_statistic}{the type of test-statistic}

\item{pval_threshold}{the threshold a pvalue must cross (be lesser than)
in order for the test-statistic to be observed. Only one of
\code{pval_threshold} or \code{threshold} must be specified, and the other one
must be set to NULL.}

\item{threshold}{an alternative way to specifiy a threshold. If a
z, t, or correlation then selection rule is assumed to be
\code{abs(x) > threshold}. If F or chisq then the selection rule
is assumed to be \code{x > threshold}. Takes precedence to
\code{pval_threshold}. Only one of \code{pval_threshold} or
\code{threshold} must be specified, and the other one
must be set to NULL.}

\item{df1}{the degrees of freedom for t and chisq statistics. The sample
size for Pearson's correlation. The numerator degrees of freedom for an
F statistic.}

\item{df2}{the denominator degrees of freedom for an F statistic.}
}
\description{
Converts p-value threshold to test-statistic threshold, or
the other way around.
}
