% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{two_sided_t}
\alias{two_sided_t}
\title{A helper function for computing the selection probability of a t test-statistic}
\usage{
two_sided_t(threshold, df, ncp)
}
\arguments{
\item{threshold}{the threhold that must be crossed (in absolute value).}

\item{df}{the degrees of freedom.}

\item{the}{non-centrality parameter.}
}
\description{
A helper function for computing the selection probability of a t test-statistic
}
