% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson-correlation.R
\name{z_to_corr}
\alias{z_to_corr}
\title{Converts z-statstic to Pearson's correlation via Fisher's Transfomration}
\usage{
z_to_corr(z_stat, samp_size)
}
\arguments{
\item{z_stat}{the z-statistic}

\item{samp_size}{the size of the sample the z statistic was computed from}
}
\description{
Should be used to transform back from z-statistic scale to correlation scale
}
